(GH-11) Remove requirement for chocolatey/chocolatey
Now that there are two Chocolatey modules, an approved one at
chocolatey/chocolatey and a supported puppetlabs/chocolatey module,
support either of them being installed. Move the requirement for
the dependency to the readme as there is no way to set up a
dependency in such a way that one of two are required.